+ Result is shown properly + Lie scale is added + Results are calculated using float + Registrarion bug is fixed

+++ MyWebApi/MyWebApi/Entities/UserActionEntities/RecalculationResult.cs
@@ -5,6 +5,6 @@ namespace WebApi.Entities.UserActionEntities
     public class RecalculationResult
     {
         public OceanStats Stats { get; set; }
-        public int TestResult { get; set; }
+        public float TestResult { get; set; }
     }
 }
